requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-ckeditor==6.7.0
pillow==10.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.5.0
psycopg[binary,pool]==3.1.12